Fixed logic error in max records, bumped version.
Previously, there was a clause that would raise an error in
validation in the event maximum_record was greater than
start_record in a query. This was based off the misunderstanding
that maximum_record is the upper limit rather than a quantity.

I have fixed this error. This variable will be refactored in
2.0.0.