Added optional version support.
It is now no longer required to include the version when
creating an SRU configuration. Some servers require it,
so I've added that section to the drivers.

This fix will likely be removed along with drivers being the
standard way to parse an explainResponse. It will be replaced
with an error hint.