Refactored public interface.
Before, the public interface for this library required importing
SRUUtil, creating an SRUConfiguration object, creating a Query
object, constructing the request, creating a session with
requests.session, sending the request, and parsing the response.
This is a cumbersome process.

I've refactored the application so that you simply have to
create an SRUQueryer with a server URL, and can then run
the search_retrieve() function directly on the queryer which
will return the content. Two steps versus 5-7, and much more
understandable as well.

This will be a part of 2.0.0, and its what the interface
should have been to start with.