Refactored available schemas, fixed default validation.
I refactored the schemas available to include the identifier for
each schema name, instead of including them as seperate entities.
This fixed a bug I didn't know about, which is that you can't
use the identifier as the schema name in the URL.

I also fixed an issue where an extra '=' sign is added in queries
with only a search term. This was causing errors.

I also realized I had forgotten to add the validator for
the defaults! This meant that I could set a bad record schema as
the default, and it goes through just fine.

Lastly, I added the '=' sign as the default relation IF the user
don't provide one and the explainResponse doesn't return it. This
is because it will be difficult for the user to determine this default
and it's directly stated in the documentation.