Added option to disable validation.
I added an option to disable validation when getting a request
object or sending a search_retrieve query. This allows you to send
or examine a searchRetrieve request that the program is failing.

This would be useful for debugging, or also disabling validation
for queries that someone suspects the library is unneccesarily
failing.