LIBILS-916. Allow for no Record Schema info.
Explain response parsing failed when the explain response did not
carry the record schema info. I had not planned for that to be
empty. So, I have made the neccesary changes so that it will
pass.

I also modified the test files so that they use os.path, which
means the unittests should work properly on Windows. There may
yet be some missing.

https://umd-dit.atlassian.net/browse/LIBILS-916